+++ CHANGELOG.md
@@ -6,8 +6,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - 2019-03-12
+## [0.4.0] - 2019-03-13
 ### Added
+- `android_hardware.Sensor.*` constants
+- `android_hardware.SensorManager.*` constants
+- `android_location.Location.*` constants
 - `android_os.*.*` constants
 
 ## [0.3.7] - 2019-03-12